requirements.txt
confusable-homoglyphs==3.2.0
dj-database-url==0.5.0
Django==1.11
django-bootstrap3==11.1.0
django-heroku==0.3.1
django-registration==2.4.1
django-tinymce==2.8.0
djangorestframework==3.10.3
gunicorn==19.9.0
Pillow==6.1.0
psycopg2==2.8.3
python-decouple==3.1
pytz==2019.2
sqlparse==0.3.0
whitenoise==4.1.3
yapf==0.28.0